lsm303d split out header and main
