+++ src/drivers/imu/lsm303d/CMakeLists.txt
@@ -37,8 +37,10 @@ px4_add_module(
 	COMPILE_FLAGS
 		-Wno-cast-align # TODO: fix and enable
 	SRCS
-		lsm303d.cpp
+		lsm303d_main.cpp
+		LSM303D.cpp
 	DEPENDS
 		drivers_accelerometer
 		drivers_magnetometer
+		px4_work_queue
 	)
